Blink at 7.7 KHz
This commit was able to build, load and run on the Nucleo F446RE.
The LED appears solid, but is blinking at 7.7 KHz (measured w/scope).

Unfortunately, the write(reg, val) procedures are not inlined.
They are very inefficient: ~14 strd instructions before each write()
and write() is VERY long (many calls to setMask() and/or masked().